requirements.txt
# Dependências diretas do ArrobaDigital (compatível com Python 3.9+).
# Evite `pip freeze` de ambientes Python 3.10+ aqui — quebra em 3.9 (ex.: contourpy, filelock).

opencv-python>=4.8.0,<5
PyYAML>=6.0,<7
requests>=2.28.0,<3
python-dotenv>=1.0.0,<2
imageio>=2.31.0,<3
psycopg2-binary>=2.9.9,<3
ultralytics>=8.3.0,<9

# Dev / testes
pytest>=7.4,<9
numpy>=1.24,<3   # já vem por ultralytics, fixo aqui por garantia dos testes

